voice-agent/requirements.txt
livekit-agents
livekit-plugins-google
livekit-plugins-silero
livekit-plugins-cartesia
livekit-plugins-assemblyai
python-dotenv
